reduce rs files to 1
